services/bot/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
websockets==12.0
pydantic==2.9.2
loguru==0.7.2
fastapi==0.115.5
uvicorn[standard]==0.32.1

# DB
psycopg[binary]==3.2.1
SQLAlchemy==2.0.47
alembic==1.18.4

# Cache/queue
redis==5.0.8
prometheus-client==0.20.0
